Drive controller generation and interpretation of trajectories is now working
